+++ main/sensors/models.py
@@ -17,6 +17,7 @@ class Sensor(models.Model):
         ACTIVE = "ACTIVE", "Active"
         INACTIVE = "INACTIVE", "Inactive"
         PENDING = "PENDING", "Pending"
+        DELETED = "DELETED", "Deleted"
 
     id = models.UUIDField(primary_key=True, default=uuid.uuid4, editable=False)
     operator = models.ForeignKey(
@@ -47,6 +48,7 @@ class Sensor(models.Model):
 
     status = models.CharField(
         max_length=20,
+        db_index=True,
         choices=SensorStatus.choices,
         default=SensorStatus.PENDING,
     )